+++ src/main/java/atsys/backtesting/engine/components/TickData.java
@@ -1,11 +1,12 @@
 package atsys.backtesting.engine.components;
 
 import atsys.backtesting.engine.components.asset.Instrument;
+import atsys.utils.Decimal;
 
 import java.time.Instant;
 
 public interface TickData {
     Instant getTickTimestamp();
-    double getLastTradedPrice();
+    Decimal getLastTradedPrice();
     Instrument getInstrument();
 }
