+++ src/main/java/atsys/backtesting/impl/components/SimpleTickData.java
@@ -2,6 +2,7 @@ package atsys.backtesting.impl.components;
 
 import atsys.backtesting.engine.components.TickData;
 import atsys.backtesting.engine.components.asset.Instrument;
+import atsys.utils.Decimal;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -12,7 +13,7 @@ import java.time.Instant;
 public class SimpleTickData implements TickData {
     private Instant tickTimestamp;
     private Instant lastTradedTime;
-    private double lastTradedQuantity;
-    private double lastTradedPrice;
+    private int lastTradedQuantity;
+    private Decimal lastTradedPrice;
     private Instrument instrument;
 }
